Create one simple schema
